refactor(userspace/plugin): rename source files
